#36 Use Map instead of MultiValueMap as options type
During the last refactorings we wondered why MultiValueMaps are used for options.
The original developer has no clue, why he used it.

First we encountered the problem that the value generic type was untyped
Object, which is bad due to type confusion bugs. So we changed it to
String, since all values are somehow converted to Strings in the end (its HTTP).

This resulted in a compile error, when adding these options to the request
body because of type erasure problems (Type<String> is not the same as
Type<Object>).

So we changed the addition of options to the request body by looping over the
options and add them explicitly. This change resulted in a runtime error on
HTTP request:

  org.springframework.http.converter.HttpMessageNotWritableException:
  Could not write request: no suitable HttpMessageConverter found for request
  type [java.util.ArrayList]

The reason for this is that values in a MultiValueMap may be a single value or
multiple values (which will be a List of values). Before our change this worked
by accident because nobody added multiple values. That's my assumption

Since we do not need multiple values under the same key I change the API to
a simple Map<String,String>.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>