Enable Dependabot for Wwrokflow updates
We want to be notified if

- used GitHub actions and
- used baes images in the workflow

are outdated.

We are not sure if it will work. This is a first try.

The docker ecosystem is configured globaly because we may use
containers outside workflows and we hope that Dependabot searches
recursive. We are also not sure if Dependabot recognizes the image
tag in the workflow YAML as Docker dependency.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>