#23 Remove BaseModel since it breaks equals/hashCode contract
Problem of equals and inheritance:
As described in this[1] blog post the Object#equals() requires that it
fulfills the Liskov Substitution Principle (LSP). Our implemenation
breaks this contract. Even worse, it didn't work at all as descibed
in issue 23[2].

Since the BaseModel class does not have any properties, we can remove it.

This patch removes this obsolete class. Also it makes all model classes
final and all properties private. This is an implicit requirement of the
contract of hashCode() to avoid memory leaks in collections. (See linked
blog post on Artima for more details.)

Actually objects must be immutable -- all fields final -- to guaruantee
a stable equal/hashCode behaviour for the whole lifetime of the objects.
But this must be further investigated, if this is possible. For now we
ignore this warning in the test code.

1: https://www.artima.com/articles/how-to-write-an-equality-method-in-java
2: https://github.com/secureCodeBox/defectdojo-client-java/issues/23

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>