Move Public API to top of Class
The structure of code should be ordered, so that the more abstract
API is at the top and the more nitty gritty details are below. This
has the advantage that one interested in an quick overview need not
to scroll up and down. You only need to scroll down for details.

Also add some null-checks and missing API doc.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>